+++ nix/shell.nix
@@ -1,0 +1,12 @@
+{
+  mkShell,
+  go,
+  gomod2nix,
+}:
+mkShell {
+  name = "lazyjira-dev";
+  packages = [
+    go
+    gomod2nix
+  ];
+}