+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- Nix: Create a standalone `lazyjira` package and shell for non flake user.
+
 ## [2.19.0] - 2026-06-22
 
 ### Added